doxygen; ctest; remove modmaker
