+++ CMakeLists.txt
@@ -77,3 +77,6 @@ add_custom_command(
 )
 
 add_subdirectory(src)
+
+enable_testing()
+add_subdirectory(test)
