--- Fk/ModMaker/qmldir
@@ -1,2 +1,0 @@
-module Fk.ModMaker
-ModMaker 1.0 main.qml
