+++ src/freekill.h
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef _FREEKILL_H
+#define _FREEKILL_H
+
+int freekill_main(int argc, char **argv);
+
+#endif // _FREEKILL_H
